Support RPM version numbers in NIM package provider
This commit adds a call to `nimclient -o showres` prior to installing
a package; this is necessary so that we can determine whether the
package is an installp package or an RPM package.  This is
unfortunately necessary because the installation command for specific
versions of an RPM differs very slightly from the command for installing
a specific installp/BFF package.

We also have to do an unfortunate amount of regex-based parsing of
the output from that command, because I have been unable to find
a command that provides concise, easily parseable output *and* shows
information about both RPMs and BFFs available on a NIM lpp source.