Remove single quotes from `fileset` CLI arg in nim provider
It seems that something must have changed in puppet since the time
that the nim provider was originally submitted.  I believe that
puppet may be handling the intepretation of quotes and other special
characters in command line arguments slightly differently than it
used to; The single quotes in the `fileset` argument to `nimclient`
were not being interpreted by the shell, so they were being passed
through to the underlying package manager as part of the package
name... and since very few packages actually have surrounding single
quotes as part of their package name, this was unsurprisingly failing.

Since the current implementations of this CLI execution stuff in
Puppet seem to be handling the grouping of the individual CLI argument
strings more strictly, the provider seems to work fine without the
single quotes (even if the value we're passing to the `filesets`
argument contains a space).