FAILED: test the new native_header feature in parsedfile
This approach is flawed.

The fixture in and of itself is fine. It cannot be fed to the parsedfile
provider itself, though. The provider is basically abstract and must be
inherited to work. The inheriting provider must populate @record_types
using the record_line function. Can we stub parsedfile thus?