Fix 'ensure=>absent' for rpms
Prior to this commit, an `ensure=>absent` for an rpm package
via the nim provider would result in a call to `installp -gu`,
which is actually only capable of uninstalling bff/installp
packages and not rpms.  This commit adds a call to `lslpp -qLc`
so that we can determine which package type we're uninstalling,
and then branches to call the appropriate command (`installp -gu`
or `rpm -e`) accordingly.