Better detection of version string for NIM package provider
Prior to this commit, if you attempted to install an RPM
from a NIM source using `ensure=>present`, and there were
more than one version of that package available, the install
command would fail because AIX's rpm system doesn't automatically
default to the newest version if it finds more than one choice.
It simply fails with an error message about there being more than
one version available.

This commit adds some logic for detecting the case where more than
one version is available, and if so, it determines the newest
version number and adds that to the installation command.  Thus,
`ensure=>present` can succeed even if there are multiple versions
of an RPM package available.