Explicitly fail on RPM downgrades in the AIX package provider
Attempted downgrades of RPM packages via the NIM package provider
do nothing because the package is superseded by a newer
version, but still don't exit non-zero. The only way to downgrade is
to force the installation, which requires us to ignore dependencies. As
this isn't palatable, we now explicitly fail if we see that the
package was superseded by a newer version.