FIXUP: make this testing approach work after all
This should not be used for real, but I'll add it to history for now for
reference. Instead we'll do what Jeff proposed: "the specs should then
exercise how we expect subclasses to take advantage of the behavior
provided by the base class."

So basically, it turned out we can even mock ParsedFile to know some record
types of its own. Meh.