WIP - Provide an example of the header handling behavior
without this patch there isn't really a good example of how the parsed
file base class should be used to handle custom headers in files.  This
is a problem because it is not clear how to write a provider that
handles custom headers like those written by Vixie cron.

This patch addresses the problem by wiring up an example class that
descends from the base ParsedFile class.  The intent of this example
provider is two fold.  First, it should parse a fake file with a custom
header.  Second, it should manage resources inside of this file without
duplicating the custom header over and over again.

The example currently fails.  The next steps are to wire up the example
provider in a way that it actually flushes the resources back to a fake
file so that we can validate the expected contents of that file object.