(AIX) Allow non-root users to introspect services
Prior to this commit, non-root users were not able to introspect
running services via the SRC service provider on AIX. The underlying
problem is the `commands` portion of the provider DSL was/is aslo
effectively a `confines`. The SRC service provider has many commands
that it uses, but non-root users may only execute about half of them.
This constraint is sensible for starting and stopping services; however,
listing status of services should be able to be done for non-root users,
and this provides feature parity for other platforms like Linux. This
commit alters the `commands` DSL method to the `optional_commands`
method so that the appropriate metaprogramming still happends, but
removes the `confines` portion implicit to `commands`, which results
in the SRC service provider being deemed suitable for non-root users
on AIX.