(AIX) Create acceptance tests for src provider
Prior to this commit, the acceptance tests for AIX's src service
provider were barely a shell of a test. This commit tests that the
service provider can
 * find a specific service
 * start the service
 * stop the service
 * enable the service
 * disable the service

The tests were also refactored to use the puppet_acceptance `teardown`
portion of the DSL so that if a test step fails, the temporary service
will be removed prior to the next test run, especially important because
the setup section will fail if the service exists already as AIX's
`mkssys` does not have a way to override the state of an existing
subsystem (service).