Explicitly reference /usr/bin/grep
GNU grep may not support the `-p` option that AIX's grep does, and
that's critical for this provider.  This commit changes the CLI
command to explicitly reference `/usr/bin/grep` to avoid any
potential issue where the user has modified the path in a way that
might make GNU grep the default.