Revert "Merge branch 'jeffmccune-yard_gemfile'"
This reverts commit 09f097e73917b43336c9dd55dc90474e495f32c4, reversing
changes made to 8f105246bc75c7cb7c77a4e2a14cb5ae211ea055.

This is being reverted because the yard gem cannot be installed properly
on Windows.

Talking to @joshcooper, we're going to revert this now and figure out a
better way to handle dependencies like this that have issues on some
platforms.  Hopefully there's a way to satisfy a dependency on all
platforms by default and exclude only those we know to be problematic.