(#18494) Remove TypeError handling code used for tests
For Ruby 1.9, commit 252c884 added code in the Puppet::Transaction::Event
class to handle strings rather than a resource object that responded to []
with symbols, for the purposes of testing.

In Ruby 2.0, the TypeError message changed a bit, so the rescue didn't work as
expected.  This is now reworked so the test passes in an object that responds
appropriately.