Fix the parsedfile spec through simplification.
Jeff, you were a bit too clever here. There was some parsing logic for key/value
pairs in your example provider, but no logic for properly writing the records
back. The effect was that the provider was effectively dropping the = signs
from the non-comment lines.

Even fixed, this would need to be a quite different test, like "with
matching rules for variable whitespace" it "should normalize the
whitespace to a desirable amount". But I believe that's way too specific
for a unit test of the base provider.

This is my fix: Ignore any semantics in the example provider. Each line is
text. Seeing as the native_header feature operates on the re-assembled output
text, this is no limitation.

This approach gives us a lot of power, but on the other hand I feel that it
is not much of a good unit test, because it will pass or fail depending on
the record types on the one hand and the fixtures on the other hand. It cannot
really make good statements on the parsedfile behaviour itself. Should this
be more meta? But that would be more difficult. Will put more thought into this.
Also needs discussing.