Test the vixie header filterin in the crontab provider
The cron test was written more like the other cron specs. It's simpler than
the base parsedfile specs. Maybe we should go back and restructure them to
look a bit more like these.

Specifically, expecting the invocation of the write method with the complete
file content we want is not flexible at all. Perhaps we can do it like the
crontab spec: Call to_file and compare the returned value with appropriate
data.

The trick is to retrieve the records to send to the to_file method. The
crontab spec does this by calling parse. This may work, so look into it.