Simplify parsedfile testing
As I hoped, following the example of the crontab spec, we can save a few
of the hoops we jumped through earlier. Way to make half of today's work
obsolete ;-)

I'm a bit dissatisfied with the execution though. It's sad that subject
apparently isn't correctly defined outside the 'it "..." do' example blocks.
Therefor, the stubbing code concerning the subject itself cannot be done in
'context' scope, which would save quite some ugly copy/paste code.