(#18978) Redo acceptance test
The previous test did not work when executed through the acceptance
framework due to quoting issues. I fixed it on *nix, but it fails on
Windows due to the RUBYLIB not being set correctly.

This commit just changes the test to ensure we can authenticate the
forge's SSL cert, which is the main motivation behind #18978 anyways.
But since the module tool is not yet supported (#11276), this test is
skipped on windows.