(#7911) Moved libuser.conf to lib/puppet/util/libuser.conf
Based on internal conversations at Puppet Labs it was recommended that the libuser.conf
file be moved out of Puppet[:confdir]/provider and into the lib/puppet/util directory.

Further discussion will be had and a decision to move this elsewhere may be made, but
for the time being this is where it will stay.