(#19247) Defer to Puppet::Util::Colors for :notice, :info, and :debug colors
Not sure what we want to do about the upper 5 (:emerg, :alert, :crit,
:err, :warning), since they deviate significantly from
Puppet::Util::Colormap.

I'd like to update them to follow this convention (and then get rid of
the concept of color in #handle and just pass the log level), but not
sure which is the right set of colors.