(#18131) Only truncate `puppet describe --list` on sentences
Previously the regexp which attempts to truncate the output
of `puppet describe --list` at the end of a sentence, if it
is within the first 45 characters of the doc string, actually
truncated at the first period/full-stop character. This caused
doc strings which include dots, such as `.k5login`, to be
prematurely truncated.

This commit causes it to instead only truncate at the first
dot-space combination, which is a more reliable indicator of
the end of a sentence.

This also fixes a typo in the spec tests for this code, but
digging further into the code to figure out how to test it
effectively sent me down a rabbit-hole and I gave up.

Paired-With: Michael Hall <mph@puppetlabs.com>