(#18092) Use URI.decode on network URI query strings
The current implementation of the network device enable password parsing
is unable to handle characters like sharps and spaces in the password,
which are valid passwords. This commit adds URI decoding of the query
string to allow all uri-encodable values to be used in the enable
password.

This supercedes GH-1327.