(#19459) Accept an empty hash in create_resources() keys parameters
Without this patch, it is not possible to have the first hash passed to
create_resources() having keys with no values, because merge() then
fails assuming the variable is nil and not an empty hash.
In practice this makes the function quite unusable to define a batch of
resources with a default set of parameters (which you can define in
another hash).