(#8858) Refactor tests to use real HTTP objects
This is a backport of commit cd4bee82:

Test SSL setup code, not our stubbing of it.

This rewrites the tests to actually test code - previously, almost every
single object involved was a stub.

There shouldn't be any functional changes, other than tests that might
actually fail at some point, falling out of this.

The results are not the most awesome of tests, but they are markedly
better than before we started.