(#15561) Remove check for invalid characters in Puppet::SSL::Base
We move it to Puppet::SSL::CertificateRequest instead as these are the
only type of SSL object that the master recieves and accepts from
agents.

The reason for this is to be still able to accept certificates
containing / and other characters when making SSL connections.

At least by default the master doesn't accept posts of certificates or
keys through the REST API.