[packaging] Update mocks for rpmbuilder mock format
The Puppet Labs mocks created by rpmbuilder have assumed a new format, pl-el-*
vs el-*, in order to avoid overwriting the configurations supplied by the mock
package. This commit updates the mocks in puppet to reflect the new standard,
so that we can continue to build packages with builders created with the
rpmbuilder module. It also removes the "rc_mocks", which was deprecated in the
packaging repo some time ago and are no longer used.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>