(#15911) automatic .service suffix for the systemd provider to keep service names consistent
This patch appends '.service' as a suffix for any service name passed to it
that does not already have a dotted suffix. It is needed to allow service
definitions to use the same service naming across systemd and non-systemd
nodes, and to provide consistency on nodes that have mixed SysV and systemd
init layouts.

Upstream systemctl assumes the .service suffix, but that will only help
Fedora 19+.