(#19090) Fix user failure when group exists and gid is not specified
When a group matching the name of a user already exists on a system
and the user resource does not contain a gid parameter, the execution
of useradd fails.

Normally useradd creates a group with an identical name to the user
when -g is not provided on the command line.  In the case where the group
already exists -g must be explicitly passed to useradd.