(#11331) Add support for 'import' in egrammar.
This adds support for 'import' expressions in the egrammar parser.
The support handles import as a special kind of function call (except
for the keyword 'import') it is indistinguishable from a non
parenthesized call and it is difficult to deal with as an expression
otherwise.

In this implementation, it is not the parsers responsibility to nest
parsing, it simply returns an ImportExpression model element.