(#19090) Fixup failing windows spec tests
Without this patch applied the spec examples are failing on windows.
The root cause of this problem is that the Etc module does not respond
to the `gid` method.

This patch addresses the problem by extending the rescue block to catch
NoMethodError exceptions in addition to the ArgumentError exceptions
that were already being caught.

References: [Build #634 (Mar 4, 2013 10:01:31
AM)](http://links.puppetlabs.com/win_19090_spec)